implement the useValidateNickname hook
